Refact : modification컨포넌트 분리 중단
Profilesetting 컴포넌트를 사용하여 Modification 컴포넌트로공통 부분을  분리 후 프로필/상품 Modify페이지를 구현하려고 했으나
코드를 분리하는떼 시간이 너무 많이 소요되어 리팩토링할 예정입니다.
