+++ src/components/common/modification/modification.module.css
@@ -1,0 +1,106 @@
+.cont_profileSetting {
+    text-align: center;
+    width: 390px;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    margin: 0 auto;
+}
+
+.cont_Img {
+    position: relative;
+}
+
+.cont_Img.profile {
+    width: 110px;
+    height: 110px;
+}
+
+.cont_Img.product {
+    width: 322px;
+    height: 236px;
+    border-radius: 10px;
+    padding: 12px;
+}
+
+/* .img_modification {
+    border-radius: 50%;
+    width: 110px;
+    height: 110px;
+    object-fit: cover; } */
+
+.img_modification.profile {
+    border-radius: 50%;
+    width: 110px;
+    height: 110px;
+    /* object-fit: cover; */
+}
+
+.img_modification.product {
+    width: 322px;
+    height: 236px;
+    border-radius: 10px;
+    /* object-fit: cover; */
+}
+
+
+.img_uploadImg {
+    position: absolute;
+    bottom: 0;
+    right: 0;
+    cursor: pointer;
+}
+
+.tit_profileSetting {
+    padding-top: 60px;
+    font-size: 24px;
+}
+
+.txt_profileSetting {
+    font-size: 14px;
+    color: #767676;
+    margin: 20px auto 35px;
+} 
+
+.label_profileSetting {
+    display: block;
+    color: #767676;
+    text-align: left;
+    width: 322px;
+    margin: 0 auto;
+}
+
+.form_profileSetting {
+    margin-top: 50px;
+}
+
+.input_profileSetting {
+    width: 322px;
+    border-left-width:0;
+    border-right-width:0;
+    border-top-width:0;
+    border-bottom-width:1px;
+    margin: 10px 0 30px 0;
+}
+
+.btn_profileSetting {
+    display: block;
+    margin: 10px auto 20px;
+    width: 322px;
+    border: 1px solid #0f0b19;
+    border-radius: 25px;
+    padding: 10px 0px;
+    cursor: pointer;
+}
+
+.inp_file {
+    display: none;
+}
+
+.p_warning {
+    font-size: 12px;
+    width: 322px;
+    margin: -25px auto 15px;
+    text-align: left;
+    color: #EB5757;
+}
