+++ src/pages/postPage/PostDetail.jsx
@@ -5,12 +5,13 @@ import Comment from "../../components/comment/Comment";
 import PostCard from "../../components/common/card/PostCard";
 import BaseURL from "../../components/common/BaseURL";
 import style from "./postDetail.module.css";
+import Header from "../../components/common/header/Header";
 
 export default function PostDetail() {
 
     return (
         <>
-            {/* <BasicHeader/> */}
+            <Header type="post"/>
             <main>
                 <div className="">
                     {/*<PostCard/>*/}
