+++ src/pages/profilePage/ProfileModify.jsx
@@ -1,0 +1,13 @@
+import React from "react";
+import Header from "../../components/common/header/Header";
+import Modification from "../../components/common/modification/Modification";
+
+export default function ProfileModify() {
+    
+    return (
+        <>
+            {/* <Header type="modification"/> */}
+            <Modification type="profile" mode="modification"/>
+        </>
+    )
+}
