+++ src/routes/Router.jsx
@@ -11,6 +11,8 @@ import Page404 from "../pages/Page404";
 import Signup from "../pages/loginPage/Signup";
 import PostDetail from "../pages/postPage/PostDetail";
 import FollowersFollowing from "../pages/followListPage/FollowersFollowing";
+import ProfileModify from "../pages/profilePage/ProfileModify";
+import ProductModify from "../pages/productPage/ProductModify";
 
 export default function Router() {
     return (
@@ -23,6 +25,9 @@ export default function Router() {
                 <Route path="/signup" element={<Signup />} />
                 <Route path="/upload" element={<Upload type = "upload"/>} />            
                 <Route path="/profile/:accountname" element={<Profile />} />           
+                <Route path="/profilemodify" element={<ProfileModify/>} />  
+                <Route path="/productmodify" element={<ProductModify/>} />            
+                <Route path="/profile" element={<Profile />} />           
                 <Route path="/profilesetting" element={<ProfileSetting />} />
                 <Route path="/profile/:accountname/:followtype" element={<FollowersFollowing />} />        
                 <Route path="/post/:postid" element={<PostDetail />} />
